Split MID-360 tooling by responsibility
